Refactor Slider Widget with abstract class
- Create abstract class d-slider, d-slider-range, d-slider-track,
d-slider-track-selected
- Move all load slider to only load slider method
- Create theme class for d-slider, d-slider-range, d-slider-track,
d-slider-track-selected
