+++ src/main/typescript/wcardinal/ui/d-slider-track-selected-horizontal.ts
@@ -3,16 +3,18 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { DButton, DButtonOptions, DThemeButton } from "./d-button";
+import {
+	DSliderTrackSelected, DSliderTrackSelectedOptions, DThemeSliderTrackSelected
+} from "./d-slider-track-selected";
 
 export interface DSliderTrackSelectedHorizontalOptions<
 	VALUE = unknown,
 	THEME extends DThemeSliderTrackSelectedHorizontal = DThemeSliderTrackSelectedHorizontal
-> extends DButtonOptions<VALUE, THEME> {
+> extends DSliderTrackSelectedOptions<VALUE, THEME> {
 
 }
 
-export interface DThemeSliderTrackSelectedHorizontal extends DThemeButton {
+export interface DThemeSliderTrackSelectedHorizontal extends DThemeSliderTrackSelected {
 
 }
 
@@ -21,7 +23,7 @@ export class DSliderTrackSelectedHorizontal<
 	THEME extends DThemeSliderTrackSelectedHorizontal = DThemeSliderTrackSelectedHorizontal,
 	OPTIONS extends DSliderTrackSelectedHorizontalOptions<VALUE, THEME>
 	= DSliderTrackSelectedHorizontalOptions<VALUE, THEME>
-> extends DButton<VALUE, THEME, OPTIONS> {
+> extends DSliderTrackSelected<VALUE, THEME, OPTIONS> {
 	protected init( options?: OPTIONS ) {
 		super.init( options );
 	}
