+++ src/main/typescript/wcardinal/ui/d-slider-track-selected-vertical.ts
@@ -3,16 +3,18 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { DButton, DButtonOptions, DThemeButton } from "./d-button";
+import {
+	DSliderTrackSelected, DSliderTrackSelectedOptions, DThemeSliderTrackSelected
+} from "./d-slider-track-selected";
 
 export interface DSliderTrackSelectedVerticalOptions<
 	VALUE = unknown,
 	THEME extends DThemeSliderTrackSelectedVertical = DThemeSliderTrackSelectedVertical
-> extends DButtonOptions<VALUE, THEME> {
+> extends DSliderTrackSelectedOptions<VALUE, THEME> {
 
 }
 
-export interface DThemeSliderTrackSelectedVertical extends DThemeButton {
+export interface DThemeSliderTrackSelectedVertical extends DThemeSliderTrackSelected {
 
 }
 
@@ -20,7 +22,7 @@ export class DSliderTrackSelectedVertical<
 	VALUE = unknown,
 	THEME extends DThemeSliderTrackSelectedVertical = DThemeSliderTrackSelectedVertical,
 	OPTIONS extends DSliderTrackSelectedVerticalOptions<VALUE, THEME> = DSliderTrackSelectedVerticalOptions<VALUE, THEME>
-> extends DButton<VALUE, THEME, OPTIONS> {
+> extends DSliderTrackSelected<VALUE, THEME, OPTIONS> {
 	protected init( options?: OPTIONS ) {
 		super.init( options );
 	}
