+++ src/main/typescript/wcardinal/ui/theme/white/d-theme-white-slider-horizontal.ts
@@ -3,22 +3,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { DBaseInteractive } from "../../d-base-interactive";
-import { DCoordinateSize } from "../../d-coordinate";
-import { DThemeWhiteBase } from "./d-theme-white-base";
+import { DThemeWhiteSlider } from "./d-theme-white-slider";
 
-export class DThemeWhiteSliderHorizontal extends DThemeWhiteBase implements DThemeWhiteSliderHorizontal {
-
-	getHeight(): DCoordinateSize {
-		return "auto";
-	}
-
-	getWidth(): DCoordinateSize {
-		return "auto";
-	}
-
-	getInteractive(): DBaseInteractive {
-		return DBaseInteractive.CHILDREN;
-	}
+export class DThemeWhiteSliderHorizontal extends DThemeWhiteSlider implements DThemeWhiteSliderHorizontal {
 
 }
