+++ src/main/typescript/wcardinal/ui/theme/white/d-theme-white-slider-range-vertical.ts
@@ -3,12 +3,10 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { DBaseInteractive } from "../../d-base-interactive";
-import { DBaseState } from "../../d-base-state";
 import { DCoordinateSize } from "../../d-coordinate";
-import { DThemeWhiteBase } from "./d-theme-white-base";
+import { DThemeWhiteSliderRange } from "./d-theme-white-slider-range";
 
-export class DThemeWhiteSliderRangeVertical extends DThemeWhiteBase implements DThemeWhiteSliderRangeVertical {
+export class DThemeWhiteSliderRangeVertical extends DThemeWhiteSliderRange implements DThemeWhiteSliderRangeVertical {
 
 	getHeight(): DCoordinateSize {
 		return 322;
@@ -18,75 +16,4 @@ export class DThemeWhiteSliderRangeVertical extends DThemeWhiteBase implements D
 		return 35;
 	}
 
-	getInteractive(): DBaseInteractive {
-		return DBaseInteractive.CHILDREN;
-	}
-
-	getPadding() {
-		return 0;
-	}
-
-	getPaddingLeft(): number {
-		return 0;
-	}
-
-	getPaddingRight(): number {
-		return 0;
-	}
-
-	getPaddingTop(): number {
-		return 0;
-	}
-
-	getPaddingBottom(): number {
-		return 0;
-	}
-
-	getMargin(): number {
-		return 0;
-	}
-
-	getMarginLeft(): number {
-		return 0;
-	}
-
-	getMarginRight(): number {
-		return 0;
-	}
-
-	getMarginTop(): number {
-		return 0;
-	}
-
-	getMarginBottom(): number {
-		return 0;
-	}
-
-	getBorderAlpha( state: DBaseState ): number {
-		return 0;
-	}
-
-	getBorderWidth( state: DBaseState ): number {
-		return 0;
-	}
-
-	getBorderAlign( state: DBaseState ): number {
-		return 0;
-	}
-
-	getOutlineAlpha( state: DBaseState ): number {
-		return 0;
-	}
-
-	getOutlineWidth( state: DBaseState ): number {
-		return 0;
-	}
-
-	getOutlineOffset( state: DBaseState ): number {
-		return 0;
-	}
-
-	getOutlineAlign( state: DBaseState ): number {
-		return 0;
-	}
 }
