+++ src/main/typescript/wcardinal/ui/theme/white/d-theme-white-slider-track-vertical.ts
@@ -3,23 +3,10 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { DBaseState } from "../../d-base-state";
-import { DCoordinatePosition, DCoordinateSize } from "../../d-coordinate";
-import { DThemeWhiteButton } from "./d-theme-white-button";
+import { DCoordinateSize } from "../../d-coordinate";
+import { DThemeWhiteSliderTrack } from "./d-theme-white-slider-track";
 
-export class DThemeWhiteSliderTrackVertical extends DThemeWhiteButton implements DThemeWhiteSliderTrackVertical {
-
-	getBackgroundColor( state: DBaseState ): number | null {
-		return 0xCCCCCC;
-	}
-
-	getBorderColor( state: DBaseState ): number | null {
-		return null;
-	}
-
-	getColor( state: DBaseState ): number {
-		return 0x5F5F5F;
-	}
+export class DThemeWhiteSliderTrackVertical extends DThemeWhiteSliderTrack implements DThemeWhiteSliderTrackVertical {
 
 	getHeight(): DCoordinateSize {
 		return 300;
@@ -27,10 +14,6 @@ export class DThemeWhiteSliderTrackVertical extends DThemeWhiteButton implements
 
 	getWidth(): DCoordinateSize {
 		return 5;
-	}
-
-	getX(): DCoordinatePosition {
-		return "CENTER";
 	}
 
 }
