+++ src/main/typescript/wcardinal/ui/theme/white/index.ts
@@ -134,6 +134,22 @@ export * from "./d-theme-white-scroll-bar-thumb";
 export * from "./d-theme-white-scroll-bar";
 export * from "./d-theme-white-select-multiple";
 export * from "./d-theme-white-select";
+export * from "./d-theme-white-slider";
+export * from "./d-theme-white-slider-horizontal";
+export * from "./d-theme-white-slider-max";
+export * from "./d-theme-white-slider-min";
+export * from "./d-theme-white-slider-range";
+export * from "./d-theme-white-slider-range-horizontal";
+export * from "./d-theme-white-slider-range-vertical";
+export * from "./d-theme-white-slider-thumb";
+export * from "./d-theme-white-slider-track";
+export * from "./d-theme-white-slider-track-horizontal";
+export * from "./d-theme-white-slider-track-selected";
+export * from "./d-theme-white-slider-track-selected-horizontal";
+export * from "./d-theme-white-slider-track-selected-vertical";
+export * from "./d-theme-white-slider-track-vertical";
+export * from "./d-theme-white-slider-value";
+export * from "./d-theme-white-slider-vertical";
 export * from "./d-theme-white-table-body-cell-button";
 export * from "./d-theme-white-table-body-cell-check";
 export * from "./d-theme-white-table-body-cell-color";
