+++ src/main/typescript/wcardinal/ui/theme/white/load/index.ts
@@ -65,18 +65,7 @@ export * from "./load-theme-white-picker-time";
 export * from "./load-theme-white-scroll-bar";
 export * from "./load-theme-white-select";
 export * from "./load-theme-white-shape-action-value";
-export * from "./load-theme-white-slider-horizontal";
-export * from "./load-theme-white-slider-track-horizontal";
-export * from "./load-theme-white-slider-track-selected-horizontal";
-export * from "./load-theme-white-slider-track-selected-vertical";
-export * from "./load-theme-white-slider-track-vertical";
-export * from "./load-theme-white-slider-thumb";
-export * from "./load-theme-white-slider-max";
-export * from "./load-theme-white-slider-min";
-export * from "./load-theme-white-slider-range-horizontal";
-export * from "./load-theme-white-slider-range-vertical";
-export * from "./load-theme-white-slider-value";
-export * from "./load-theme-white-slider-vertical";
+export * from "./load-theme-white-slider";
 export * from "./load-theme-white-table";
 export * from "./load-theme-white-text";
 export * from "./load-theme-white";
