+++ src/main/typescript/wcardinal/ui/theme/white/load/load-theme-white-all.ts
@@ -29,18 +29,7 @@ import { loadThemeWhitePickerDate } from "./load-theme-white-picker-date";
 import { loadThemeWhitePickerDatetime } from "./load-theme-white-picker-datetime";
 import { loadThemeWhitePickerTime } from "./load-theme-white-picker-time";
 import { loadThemeWhiteSelect } from "./load-theme-white-select";
-import { loadThemeWhiteSliderHorizontal } from "./load-theme-white-slider-horizontal";
-import { loadThemeWhiteSliderMax } from "./load-theme-white-slider-max";
-import { loadThemeWhiteSliderMin } from "./load-theme-white-slider-min";
-import { loadThemeWhiteSliderRange } from "./load-theme-white-slider-range-horizontal";
-import { loadThemeWhiteSliderRangeVertical } from "./load-theme-white-slider-range-vertical";
-import { loadThemeWhiteSliderThumb } from "./load-theme-white-slider-thumb";
-import { loadThemeWhiteSliderTrackHorizontal } from "./load-theme-white-slider-track-horizontal";
-import { loadThemeWhiteSliderTrackSelectedHorizontal } from "./load-theme-white-slider-track-selected-horizontal";
-import { loadThemeWhiteSliderTrackSelectedVertical } from "./load-theme-white-slider-track-selected-vertical";
-import { loadThemeWhiteSliderTrackVertical } from "./load-theme-white-slider-track-vertical";
-import { loadThemeWhiteSliderValue } from "./load-theme-white-slider-value";
-import { loadThemeWhiteSliderVertical } from "./load-theme-white-slider-vertical";
+import { loadThemeWhiteSlider } from "./load-theme-white-slider";
 import { loadThemeWhiteTable } from "./load-theme-white-table";
 import { loadThemeWhiteText } from "./load-theme-white-text";
 import { loadThemeWhiteView } from "./load-theme-white-view";
@@ -75,16 +64,5 @@ export const loadThemeWhiteAll = () => {
 	loadThemeWhiteSelect();
 	loadThemeWhiteTable();
 	loadThemeWhiteText();
-	loadThemeWhiteSliderHorizontal();
-	loadThemeWhiteSliderRange();
-	loadThemeWhiteSliderTrackHorizontal();
-	loadThemeWhiteSliderTrackSelectedHorizontal();
-	loadThemeWhiteSliderThumb();
-	loadThemeWhiteSliderMax();
-	loadThemeWhiteSliderMin();
-	loadThemeWhiteSliderValue();
-	loadThemeWhiteSliderVertical();
-	loadThemeWhiteSliderRangeVertical();
-	loadThemeWhiteSliderTrackVertical();
-	loadThemeWhiteSliderTrackSelectedVertical();
+	loadThemeWhiteSlider();
 };
