--- src/main/typescript/wcardinal/ui/theme/white/load/load-theme-white-slider-max.ts
@@ -1,11 +1,0 @@
-/*
- * Copyright (C) 2019 Toshiba Corporation
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import { DThemeWhite } from "../d-theme-white";
-import { DThemeWhiteSliderMax } from "../d-theme-white-slider-max";
-
-export const loadThemeWhiteSliderMax = () => {
-	DThemeWhite.set( "DSliderMax", DThemeWhiteSliderMax );
-};
