--- src/main/typescript/wcardinal/ui/theme/white/load/load-theme-white-slider-min.ts
@@ -1,11 +1,0 @@
-/*
- * Copyright (C) 2019 Toshiba Corporation
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import { DThemeWhite } from "../d-theme-white";
-import { DThemeWhiteSliderMin } from "../d-theme-white-slider-min";
-
-export const loadThemeWhiteSliderMin = () => {
-	DThemeWhite.set( "DSliderMin", DThemeWhiteSliderMin );
-};
