--- src/main/typescript/wcardinal/ui/theme/white/load/load-theme-white-slider-range-vertical.ts
@@ -1,11 +1,0 @@
-/*
- * Copyright (C) 2019 Toshiba Corporation
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import { DThemeWhite } from "../d-theme-white";
-import { DThemeWhiteSliderRangeVertical } from "../d-theme-white-slider-range-vertical";
-
-export const loadThemeWhiteSliderRangeVertical = () => {
-	DThemeWhite.set( "DSliderRangeVertical", DThemeWhiteSliderRangeVertical );
-};
