--- src/main/typescript/wcardinal/ui/theme/white/load/load-theme-white-slider-track-vertical.ts
@@ -1,11 +1,0 @@
-/*
- * Copyright (C) 2019 Toshiba Corporation
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import { DThemeWhite } from "../d-theme-white";
-import { DThemeWhiteSliderTrackVertical } from "../d-theme-white-slider-track-vertical";
-
-export const loadThemeWhiteSliderTrackVertical = () => {
-	DThemeWhite.set( "DSliderTrackVertical", DThemeWhiteSliderTrackVertical );
-};
