--- src/main/typescript/wcardinal/ui/theme/white/load/load-theme-white-slider-value.ts
@@ -1,11 +1,0 @@
-/*
- * Copyright (C) 2019 Toshiba Corporation
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import { DThemeWhite } from "../d-theme-white";
-import { DThemeWhiteSliderValue } from "../d-theme-white-slider-value";
-
-export const loadThemeWhiteSliderValue = () => {
-	DThemeWhite.set( "DSliderValue", DThemeWhiteSliderValue );
-};
